feat: added fake database seeding
